+++ src/commands/test.ts
@@ -2,13 +2,11 @@ import { Arguments, CommandBuilder } from "yargs";
 import spinner from "../services/spinner";
 
 import { authenticate, BaseOptions, baseOptions, printError } from "../shared";
-import { validateMetadata } from "./deploy/handlers/validators";
 import handlers from "./test/handlers";
 
 export interface Options extends BaseOptions {
-  config?: string;
   "out-file": string;
-  service?: string;
+  service: string;
 }
 
 export const command = "test";
@@ -18,20 +16,14 @@ export const builder: CommandBuilder<Options, Options> = (yargs) => {
   return yargs
     .options({
       ...baseOptions,
-      config: {
-        type: "string",
-        desc: "The configuration folder of the service to test. Defaults to the service specified in the .baselime folder, if it exists.",
-        alias: "c",
-        default: ".baselime",
-      },
       "out-file": { type: "string", desc: "The file to output the results to", alias: "o", default: "baselime-snapshot.json" },
-      service: { type: "string", desc: "The service to test. This will be used to determine the service if no service is provided." },
+      service: { type: "string", desc: "The service to test. This will be used to determine the service if no service is provided.", required: true },
     })
     .example([
       [
         `
       $0 test
-      $0 test --config .baselime --out-file file.json`,
+      $0 test --out-file file.json`,
       ],
     ])
     .fail((message, err, yargs) => {
@@ -41,12 +33,9 @@ export const builder: CommandBuilder<Options, Options> = (yargs) => {
 
 export async function handler(argv: Arguments<Options>) {
   spinner.init(!!argv.quiet);
-  const { config, profile, "out-file": outFile, format, "api-key": apiKey } = argv;
-  let { service } = argv;
+  const { service, config, profile, "out-file": outFile, format, "api-key": apiKey } = argv;
   spinner.init(!!argv.quiet);
   await authenticate(profile, apiKey);
-
-  service = service || (await validateMetadata(config!)).service;
 
   await handlers.test(format!, { service, outFile });
 }