+++ src/services/api/api.ts
@@ -9,14 +9,10 @@ import events from "./paths/events";
 import cli from "./paths/cli";
 import iam from "./paths/iam";
 import aws from "./paths/aws";
-import templates from "./paths/templates";
-import diffs from "./paths/diffs";
-import infrastructure from "./paths/infrastructure";
 import reports from "./paths/reports";
 import slack from "./paths/slack";
 import keys from "./paths/keys";
 import onboarding from "./paths/onboarding";
-import explain from "./paths/explain";
 import environments from "./paths/environments";
 import rehydrate from "./paths/rehydrate";
 import markers from "./paths/markers";
@@ -34,15 +30,11 @@ export default {
   ...cli,
   ...iam,
   ...aws,
-  ...templates,
-  ...diffs,
-  ...infrastructure,
   ...reports,
   ...alertChecks,
   ...slack,
   ...keys,
   ...onboarding,
-  ...explain,
   ...environments,
   ...rehydrate,
   ...markers,