+++ CHANGELOG.md
@@ -6,7 +6,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
-
+### Deprecated
+- A few commands and all associatd code:
+  - deploy
+  - pull
+  - templates
+  - validate
 
 ## [0.0.79] 2023-09-23
 ### Added